+++ src/lib/components/Motif/GlobalIconWrapper/GlobalIconWrapper.module.scss
@@ -4,6 +4,10 @@
   display: inline-flex;
   align-items: center;
 
+  > :not(svg) {
+    font-size: 1em;
+  }
+
   > svg {
     width: 1em;
     height: 1em;
