+++ src/lib/components/Motif/GlobalIconWrapper/GlobalIconWrapper.test.tsx
@@ -1,31 +1,7 @@
 import "@testing-library/jest-dom";
-import { render, RenderResult } from "@testing-library/react";
+import { render } from "@testing-library/react";
 import GlobalIconWrapper from "./GlobalIconWrapper";
-import { ReactElement } from "react";
-
-export type ComponentWithIconRender = (icon: string | ReactElement) => RenderResult;
-
-export const runIconPropTest = (renderIcon: ComponentWithIconRender, className?: string) => {
-  const iconName = "home";
-
-  const testStringIcon = () => {
-    const { getByText } = renderIcon(iconName);
-    const iconElement = getByText(iconName);
-    expect(iconElement.tagName.toLowerCase()).toBe("span");
-    className && expect(iconElement).toHaveClass(className);
-  };
-
-  const testReactElementIcon = () => {
-    const { getByText } = renderIcon(<span>icon-element</span>);
-    const externalIcon = getByText("icon-element");
-    expect(externalIcon.tagName.toLowerCase()).toBe("span");
-    expect(externalIcon.parentElement).toHaveClass("Root");
-    className && expect(externalIcon.parentElement).toHaveClass(className);
-  };
-
-  testStringIcon();
-  testReactElementIcon();
-};
+import { runIconPropTest } from "../../../../utils/testUtils";
 
 describe("GlobalIconWrapper", () => {
   it("should render the main icon given in the icon prop", () => {
